# BUG: Fixes incorrect behaviour in market open and close