Fixes schedule_function's handling of after_open and before_close dates. 
We now look at *this* period start and period end, instead of the next period start and period end. The next period start and period end are implicitly creating state(when we call a rule for the first time, when its `_next_period_start` is None, and then again when its `_next_period_start `has a value). 

As a result we were seeing the following impure behaviour:
```python
In [33]: bc.should_trigger(test_close_time2, env)
Out[33]: True

In [34]: bc.should_trigger(test_close_time2, env)
Out[34]: False

In [35]: bc.should_trigger(test_close_time2, env)
Out[35]: False
```
because of the messed up state, we weren't correctly keeping track of where we were, which led to messed up behaviour with `schedule_function`. We avoid that here by simply keeping track of this period start and end, and updating the period if the date we're on is not the same as the cached date. In tests, we now only create one before_close and one after_open object, so that we can catch any regressions that affect their internal state, instead of creating fresh objects in each subtest that have a good internal state.

There is some performance impact, but we're still doing OK. Here's the code I used to test perf:
**this branch**
```python
In [11]: tdays = env.trading_days[2000:3000]

In [12]: ao = time_rules.market_open(minutes=30)

In [13]: def test_trigger(tdays, rule):
   ....:     for tday in tdays:
   ....:         mins = env.market_minutes_for_day(tday)
   ....:         for min in mins:
   ....:             rule.should_trigger(min, env)
   ....:             

In [14]: %timeit test_trigger(tdays, ao)
1 loops, best of 3: 1.13 s per loop

In [15]: bc = time_rules.market_close(minutes=30)

In [16]: %timeit test_trigger(tdays, ao)
1 loops, best of 3: 1.13 s per loop

In [17]: %timeit test_trigger(tdays, bc)
1 loops, best of 3: 1.07 s per loop
```

**lazy-mainline**
```python
In [1]: def test_trigger(tdays, rule):
    for tday in tdays:
        mins = env.market_minutes_for_day(tday)
        for min in mins:
            rule.should_trigger(min, env)
   ...:             

In [2]: ao = time_rules.market_open(minutes=30)

In [3]: tdays = env.trading_days[2000:3000]

In [4]: %timeit test_trigger(tdays, ao)
1 loops, best of 3: 838 ms per loop

In [5]: bc = time_rules.market_close(minutes=30)

In [6]: %timeit test_trigger(tdays, bc)
1 loops, best of 3: 730 ms per loop
```

**master**
```python
In [1]: def test_trigger(tdays, rule):
    for tday in tdays:
        mins = env.market_minutes_for_day(tday)
        for min in mins:
            rule.should_trigger(min, env)
   ...:             

In [2]: ao = time_rules.market_open(minutes=30)

In [4]: tdays = env.trading_days[2000:3000]

In [5]: %timeit test_trigger(tdays, ao)
1 loops, best of 3: 2.55 s per loop

In [6]: bc = time_rules.market_close(minutes=30)

In [7]: %timeit test_trigger(tdays, bc)
1 loops, best of 3: 2.51 s per loop
```