@abhijeetkalyan 2016-03-08
@jbredeche updated this PR with some performance improvements. We now do a less-than-equals check to see whether we need to calculate a new date(if `dt` is gte the `period_close`, we calculate a new date). This puts us back to lazy-mainline level perf numbers. The updated perf numbers for this branch are:
```python
In [1]: tdays = env.trading_days[2000:3000]

In [2]: def test_trigger(tdays, rule):
    for tday in tdays:
        mins = env.market_minutes_for_day(tday)
        for min in mins:
            rule.should_trigger(min, env)
   ...:             

In [3]: ao = time_rules.market_open(minutes=30)

In [4]: %timeit test_trigger(tdays, ao)
1 loops, best of 3: 851 ms per loop

In [5]: bc = time_rules.market_close(minutes=30)

In [6]: %timeit test_trigger(tdays, bc)
1 loops, best of 3: 858 ms per loop

```
**However**, we're now relying explicitly on the fact that the clock only moves forward. This logic will break if we ever pass a date prior to what we've passed before, which means that these rules are no longer truly stateless - but in our world of time only marching forward, they will behave just like any other stateless rules.

Since this now depends on time only moving forward, I've also updated the tests to pass in a sorted list of random dates, when testing before close and after open.